requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
easyocr
ollama
Pillow
pdfplumber
python-docx
pydantic
